Track drafts as well
I'm of two minds about this. On the one hand, I get to save drafts to
GitHub and increase my commit count (BECAUSE THAT'S WHAT MATTERS), but
on the other I lose the pretense that blog posts appear fully-formed out
of nowhere. Bismarck's quote about laws and sausages comes to mind.

Ah well.